% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{collapse_results}
\alias{collapse_results}
\title{Collapse enrichment results to remove duplicate genes per phenotype}
\usage{
collapse_results(results, impc_df_phenotypes, user_gene_list)
}
\arguments{
\item{results}{Enrichment results}

\item{impc_df_phenotypes}{Background gene-phenotype pairs}

\item{user_gene_list}{User gene list}
}
\value{
Deduplicated, merged results
}
\description{
Collapse enrichment results to remove duplicate genes per phenotype
}
