% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{enrichIMPC}
\alias{enrichIMPC}
\title{Run full enrichment pipeline on IMPC data}
\usage{
enrichIMPC(user_gene_list, search_terms, enrichment_path)
}
\arguments{
\item{user_gene_list}{Dataframe with user gene list (column 'Gene')}

\item{search_terms}{Terms to filter phenotypes}

\item{enrichment_path}{Output path for writing results (not currently used)}
}
\description{
Run full enrichment pipeline on IMPC data
}
