% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{filter_by_user_genes}
\alias{filter_by_user_genes}
\title{Filter the IMPC dataset using a user-provided gene list}
\usage{
filter_by_user_genes(user_gene_list, impc_df_subset)
}
\arguments{
\item{user_gene_list}{Dataframe with a column 'Gene'}

\item{impc_df_subset}{Filtered IMPC dataframe}
}
\value{
Subset of rows where Gene matches user input
}
\description{
Filter the IMPC dataset using a user-provided gene list
}
