% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{filter_phenotypes}
\alias{filter_phenotypes}
\title{Filter IMPC dataframe by phenotype search terms}
\usage{
filter_phenotypes(impc_df, search_terms)
}
\arguments{
\item{impc_df}{Dataframe of IMPC results}

\item{search_terms}{Character vector of terms to search within \verb{Phenotype Hits}}
}
\value{
Filtered dataframe with only relevant phenotypes
}
\description{
Filter IMPC dataframe by phenotype search terms
}
