% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{generate_plot}
\alias{generate_plot}
\title{Create bar plot for top 20 phenotypes with gene count and p-value labels}
\usage{
generate_plot(mouse_user_genes, results)
}
\arguments{
\item{mouse_user_genes}{User genes with phenotype hits}

\item{results}{Enrichment results}
}
\description{
Create bar plot for top 20 phenotypes with gene count and p-value labels
}
