% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{run_enrichment}
\alias{run_enrichment}
\title{Perform enrichment tests per phenotype}
\usage{
run_enrichment(impc_df_phenotypes_user_input, impc_df_phenotypes)
}
\arguments{
\item{impc_df_phenotypes_user_input}{User gene-phenotype pairs}

\item{impc_df_phenotypes}{Background gene-phenotype pairs}
}
\value{
Enrichment results with p-values and adjusted p-values
}
\description{
Perform enrichment tests per phenotype
}
