% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{split_phenotypes}
\alias{split_phenotypes}
\title{Split phenotype annotations and retain shared phenotypes}
\usage{
split_phenotypes(impc_df, mouse_user_genes)
}
\arguments{
\item{impc_df}{Full IMPC dataset}

\item{mouse_user_genes}{Subset of IMPC matching user genes}
}
\value{
A list of split dataframes used in enrichment
}
\description{
Split phenotype annotations and retain shared phenotypes
}
