% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_IMPC_genes.R
\name{write_results}
\alias{write_results}
\title{Write final enrichment results to CSV}
\usage{
write_results(results, enrichment_path)
}
\arguments{
\item{results}{Enrichment results}

\item{enrichment_path}{Output path (not used but retained for compatibility)}
}
\description{
Write final enrichment results to CSV
}
